refactor(task): extract TaskCard component and enhance TaskGrid UI
- Move TaskCard component into its own file for better modularity
- Add hover spotlight and glow effects to TaskCard for improved UX
- Refactor TaskGrid to use imported TaskCard component
- Enhance Today and Upcoming sections with blurred background rings
- Add relative positioning and gradient overlays to EmptyState component
- Remove unused icon imports from TaskGrid component for cleaner code
